chore: clear clippy lints in changed crates
Quality cleanups surfaced while verifying the review fixes (clippy is not a
CI gate, so these had accumulated):

- persistence/src/meeting_ops.rs: move `mod tests` to end of file
  (clippy::items_after_test_module — a non-test fn followed the test module).
- persistence/tests/yjs_interop.rs: drop the redundant `.or_else` closure
  (clippy::unnecessary_lazy_evaluations); the whole-float→int collapse is a
  plain value.
- persistence/src/tests.rs: drop the redundant `SAMPLE_RATE as u32` casts
  (SAMPLE_RATE is already u32; clippy::unnecessary_cast).
- common/src/lib.rs: `map_or(false, …)` → `is_some_and`
  (clippy::unnecessary_map_or).

SKIP_ARCH_CHECK: lint-only edits (test-module placement, redundant cast/closure,
an idiom swap); no component boundary, interface, dependency, or cross-cutting
change.

+++ crates/persistence/src/tests.rs
@@ -883,7 +883,7 @@ fn test_decode_opus_ogg_handles_double_header_set() {
         head.push(1); // version
         head.push(1); // channels
         head.extend_from_slice(&3840u16.to_le_bytes()); // pre_skip
-        head.extend_from_slice(&(SAMPLE_RATE as u32).to_le_bytes());
+        head.extend_from_slice(&SAMPLE_RATE.to_le_bytes());
         head.extend_from_slice(&0u16.to_le_bytes()); // gain
         head.push(0); // mapping family
         pw.write_packet(head, extra_serial, PacketWriteEndInfo::EndPage, 0)
@@ -950,7 +950,7 @@ fn test_decode_opus_ogg_extra_tags_packet_skipped() {
         head.push(1);
         head.push(1);
         head.extend_from_slice(&3840u16.to_le_bytes());
-        head.extend_from_slice(&(SAMPLE_RATE as u32).to_le_bytes());
+        head.extend_from_slice(&SAMPLE_RATE.to_le_bytes());
         head.extend_from_slice(&0u16.to_le_bytes());
         head.push(0);
         pw.write_packet(head, extra_serial, PacketWriteEndInfo::EndPage, 0)
